move loc tests from TestMultiIndexBasic class
